add useTTSPlayer hook
